@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <=3.2.0"
+  sdk: ">=2.15.0 <=3.2.0"
 
 dependencies:
   animated_splash_screen: ^1.1.0
@@ -30,6 +30,7 @@ dependencies:
     git:
       url: https://github.com/nunesricardo88/gs1_barcode_parser
   http: ^1.1.0
+  infinite_scroll_pagination: ^4.0.0
   intl: ^0.18.0
   lint: ^2.0.1
   loading_overlay: ^0.3.0
