@@ -148,67 +148,6 @@ class AlwaysDisabledFocusNode extends FocusNode {
   bool get hasFocus => false;
 }
 
-//TextStyles
-const kAppBarTextStyle = TextStyle(
-  fontFamily: 'Roboto',
-  color: kPrimaryColorLight,
-);
-
-const kButtonTextStyle = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 16.0,
-  fontWeight: FontWeight.w400,
-);
-
-const kContentLabelTextStyleSemiBold = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 18.0,
-  fontWeight: FontWeight.w500,
-  color: kPrimaryColorDark,
-);
-
-const kContentLabel2TextStyle = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 18.0,
-  fontWeight: FontWeight.normal,
-  color: kPrimaryColorDark,
-);
-
-const kContentTextStyle = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 18.0,
-  fontWeight: FontWeight.normal,
-  color: kPrimaryColorDark,
-);
-
-const kTileText = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 16.0,
-  fontWeight: FontWeight.normal,
-  color: kPrimaryTextColor,
-);
-
-const kTileFieldTitle = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 16.0,
-  fontWeight: FontWeight.w500,
-  color: kPrimaryColor,
-);
-
-const kTileFieldLabel = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 14.0,
-  fontWeight: FontWeight.normal,
-  color: kSecondaryTextColor,
-);
-
-const kTileFieldContent = TextStyle(
-  fontFamily: 'Roboto',
-  fontSize: 14.0,
-  fontWeight: FontWeight.normal,
-  color: kPrimaryTextColor,
-);
-
 //Input Decorations
 const kSearchFieldInputDecoration = InputDecoration(
   labelText: 'Procurar...',
@@ -218,6 +157,20 @@ const kSearchFieldInputDecoration = InputDecoration(
   alignLabelWithHint: true,
   border: UnderlineInputBorder(),
   focusedBorder: UnderlineInputBorder(),
+);
+
+const kPickTextFieldsInputDecoration = InputDecoration(
+  contentPadding: EdgeInsets.zero,
+  border: OutlineInputBorder(
+    borderSide: BorderSide(
+      color: kPrimaryColorDark,
+    ),
+  ),
+  labelStyle: TextStyle(
+    color: kPrimaryColor,
+  ),
+  filled: true,
+  fillColor: kWhiteBackground,
 );
 
 BoxDecoration get pinPutDecoration {
